Updated ruby environment query to explicitly use ruby we just changed to.
On OS X shelling out to ruby was using the system ruby, even after the PATH env var was updated to the new ruby. This was preventing important env vars like GEM_HOME and GEM_PATH from being set properly. Calling ruby with an explicit bin path ensures that we are exectuting the right instance.